+++ src/main/resources/templates/themes/default/page-category.html
@@ -11,7 +11,7 @@
         #else
         <div class="post-lists">
             <div class="post-lists-body">
-                #for(article : articles.list)
+                #for(article : articles.rows)
                 <div class="post-list-item">
                     <div class="post-list-item-container ">
                         <div class="item-label ">
